feat: add sat conversion to send screen privacy levels based on settings
